+++ src/extdefs/optimization_defs.jl
@@ -27,7 +27,7 @@ $(TYPEDFIELDS)
     IA<:InitvalAlgorithm
 } <: AbstractModeEstimator
     optalg::ALG = ext_default(pkgext(Val(:Optimization)), Val(:DEFAULT_OPTALG))
-    pretransform::TR = PriorToGaussian()
+    pretransform::TR = PriorToNormal()
     init::IA = InitFromTarget()
     maxiters::Int64 = 1_000
     maxtime::Float64 = NaN