+++ test/measures/test_bat_pushfwd_measure.jl
@@ -99,7 +99,7 @@ using Optim
         prior = HierarchicalDistribution(f_secondary, primary_dist)
         likelihood = logfuncdensity(logdensityof(varshape(prior)(MvNormal(Diagonal(fill(1.0, totalndof(varshape(prior))))))))
         m = PosteriorMeasure(likelihood, prior)
-        hmc_samples = bat_sample(m, TransformedMCMC(mcalg = HamiltonianMC(), pretransform = PriorToGaussian(), nsteps = 10^4), context).result
+        hmc_samples = bat_sample(m, TransformedMCMC(mcalg = HamiltonianMC(), pretransform = PriorToNormal(), nsteps = 10^4), context).result
         is_samples = bat_sample(m, PriorImportanceSampler(nsamples = 10^4), context).result
         @test isapprox(mean(unshaped.(hmc_samples)), mean(unshaped.(is_samples)), rtol = 0.1)
         @test isapprox(cov(unshaped.(hmc_samples)), cov(unshaped.(is_samples)), rtol = 0.2)